% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_forecasts.R
\name{bom_forecasts}
\alias{bom_forecasts}
\title{Get daily or 3-hourly weather forecasts}
\usage{
bom_forecasts(
  geohash,
  freq = c("daily", "3-hourly"),
  raw = FALSE,
  ep = endpoint
)
}
\arguments{
\item{geohash}{A \code{character} string giving the location ID, as retrieved
from \code{\link{bom_search_station}}}

\item{freq}{A \code{character} string giving the frequency of forecasts;
one of \code{"daily"} or \code{"3-hourly"}}

\item{raw}{A \code{logical} scalar; if \code{TRUE} returns JSON output as a \code{list},
otherwise coerces output to a \code{data.frame}}

\item{ep}{A \code{character} string of the API endpoint; defaults to
\code{weatherBOM:::endpoint}}
}
\value{
A \code{data.frame} or \code{list}
}
\description{
Get daily or 3-hourly weather forecasts
}
\details{
This function retrieves weather forecasts for a specific location ID (the
geohash key as retrieved from \code{\link{bom_search_station}}). If \code{freq = "daily"},
forecasts for the next 7 days include maximum/minimum temperatures, chance
and amount of rain, UV risk category and sun protection times, and
sunrise/sunset times. If \code{freq = "3-hourly"}, forecasts for the next 2 days
include the temperature, chance and amount of rain, and wind speed and
direction. \strong{Note that all times are given in local time.}
}
\examples{
bom_forecasts("r3dp4vp")

}
