% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_location_info.R
\name{bom_location_info}
\alias{bom_location_info}
\title{Get location information based on location ID}
\usage{
bom_location_info(geohash, ep = endpoint)
}
\arguments{
\item{geohash}{A \code{character} string giving the location ID, as retrieved
from \code{\link{bom_search_station}}}

\item{ep}{A \code{character} string of the API endpoint; defaults to
\code{weatherBOM:::endpoint}}
}
\value{
A \code{list}
}
\description{
Get location information based on location ID
}
\examples{
bom_location_info("r3dp4vp")

}
