% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_observations.R
\name{bom_observations}
\alias{bom_observations}
\title{Get current weather observations}
\usage{
bom_observations(geohash, ep = endpoint)
}
\arguments{
\item{geohash}{A \code{character} string giving the location ID, as retrieved
from \code{\link{bom_search_station}}}

\item{ep}{A \code{character} string of the API endpoint; defaults to
\code{weatherBOM:::endpoint}}
}
\value{
A \code{list}
}
\description{
Get current weather observations
}
\examples{
bom_observations("r3dp4vp")

}
