% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom_search_station.R
\name{bom_search_station}
\alias{bom_search_station}
\title{Search for a weather station based on location name}
\usage{
bom_search_station(name, raw = FALSE, ep = endpoint)
}
\arguments{
\item{name}{A \code{character} string giving the location name}

\item{raw}{A \code{logical} scalar; if \code{TRUE} returns JSON output as a \code{list},
otherwise coerces output to a \code{data.frame}}

\item{ep}{A \code{character} string of the API endpoint; defaults to
\code{weatherBOM:::endpoint}}
}
\value{
A \code{data.frame} or \code{list}
}
\description{
Search for a weather station based on location name
}
\examples{
bom_search_station("Canberra")

}
