% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compass_angle.R
\name{compass2angle}
\alias{compass2angle}
\title{Convert 16-wind compass rose points to angles}
\usage{
compass2angle(x)
}
\arguments{
\item{x}{A \code{character} vector of compass points}
}
\value{
A \code{numeric} vector
}
\description{
Convert 16-wind compass rose points to angles
}
\examples{
compass2angle("SSW")

}
