% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compass_angle.R
\name{angle2compass}
\alias{angle2compass}
\title{Convert 16-wind compass rose points to angles}
\usage{
angle2compass(x)
}
\arguments{
\item{x}{A \code{numeric} vector of angles}
}
\value{
A \code{character} vector of compass points
}
\description{
Convert 16-wind compass rose points to angles
}
\examples{
angle2compass(c(180, 22.5))

}
